♻️ refactor: 트랜잭션 유스케이스에서 관리하도록 수정
